requirements.txt
# aiofiles==23.2.1
# aiogram==3.2.0
# # aiohttp==3.8.5
# aiosignal==1.3.1
# alembic==1.13.1
# annotated-types==0.6.0
# anyio==4.0.0
# argon2-cffi==23.1.0
# argon2-cffi-bindings==21.2.0
# arrow==1.3.0
# asgiref==3.7.2
# asttokens==2.2.1
# async-lru==2.0.4
# async-timeout==4.0.2
# attrs==23.1.0
# Authlib==1.2.1
# autobahn==23.1.2
# Automat==22.10.0
# azure-common==1.1.28
# azure-core==1.30.1
# azure-identity==1.15.0
# azure-mgmt-core==1.4.0
# azure-mgmt-rdbms==10.1.0
# azure-mgmt-resource==23.0.1
# azure-mgmt-subscription==3.1.1
# Babel==2.14.0
# backcall==0.2.0
bcrypt==4.0.1
# beautifulsoup4==4.12.2
# bidict==0.23.1
# bleach==6.1.0
# blinker==1.7.0
# boto3==1.33.13
# botocore==1.33.13
# Brotli==1.1.0
# cachetools==5.3.3
# certifi==2024.2.2
# cffi==1.16.0
# channels==4.0.0
# charset-normalizer==3.3.2
# click==8.1.7
# colorama==0.4.6
# comm==0.1.3
# constantly==15.1.0
# cryptography==42.0.5
daphne==4.0.0
# debugpy==1.6.7
# decorator==5.1.1
# defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.0.3
django-channels==0.7.0
django-cors-headers==4.3.1
django-environ==0.9.0
django-googledrive-storage==1.6.0
django-rest-framework==0.1.0
Django-Verify-Email==2.0.3
django-widget-tweaks==1.4.8
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-yasg==1.21.7
ecdsa==0.18.0
email_validator==2.1.1
eventlet==0.34.2
executing==1.2.0

# google-api-core==2.17.1
# google-api-python-client==2.122.0
# google-auth==2.28.2
# google-auth-httplib2==0.2.0
# google-auth-oauthlib==1.1.0
# googleapis-common-protos==1.63.0
# greenlet==3.0.3
gunicorn==21.2.0
# h11==0.14.0
# httpagentparser==1.9.5
# httplib2==0.22.0
# hyperlink==21.0.0
# idna==3.6
# importlib_metadata==7.0.2
# importlib_resources==6.3.0
# incremental==22.10.0
# inflection==0.5.1
# ipykernel==6.24.0
# ipython==8.14.0
# ipython-genutils==0.2.0
# ipywidgets==8.1.1
# isodate==0.6.1
# isoduration==20.11.0
# itsdangerous==2.1.2
# jaraco.classes==3.3.1
# jedi==0.18.2
# jeepney==0.8.0
# Jinja2==3.1.3
# jmespath==1.0.1
# json5==0.9.14
# jsonformatter==0.3.2
# jsonpointer==2.4
# jsonschema==4.19.1
# jsonschema-specifications==2023.7.1
# jupyter==1.0.0
# jupyter-console==6.6.3
# jupyter-events==0.8.0
# jupyter-lsp==2.2.0
# jupyter_client==8.3.0
# jupyter_core==5.3.1
# jupyter_server==2.9.1
# jupyter_server_terminals==0.4.4
# jupyterlab==4.0.7
# jupyterlab-pygments==0.2.2
# jupyterlab-widgets==3.0.9
# jupyterlab_server==2.25.0
# keyring==24.3.1
# ldap3==2.9.1
# magic-filter==1.0.12
# Mako==1.3.2
# markdown-it-py==3.0.0
# MarkupSafe==2.1.5
# matplotlib-inline==0.1.6
# mdurl==0.1.2
# mistune==3.0.2
# more-itertools==10.2.0
# msal==1.27.0
# msal-extensions==1.1.0
# msrest==0.7.1
# multidict==6.0.4
# nbclient==0.8.0
# nbconvert==7.9.2
# nbformat==5.9.2
# nest-asyncio==1.5.6
# notebook==7.0.6
# notebook_shim==0.2.3
# oauthlib==3.2.2
# openai==0.27.8
# overrides==7.4.0
# packaging==24.0
# pandocfilters==1.5.0
# paramiko==3.4.0
# parso==0.8.3
# passlib==1.7.4
# pickleshare==0.7.5
# pillow==10.2.0
# platformdirs==3.8.0
# portalocker==2.8.2
# prometheus-client==0.17.1
# prompt-toolkit==3.0.39
# protobuf==4.25.3
# psutil==5.9.8

psycopg2==2.9.9
psycopg2-binary==2.9.9
# pure-eval==0.2.2
# pyasn1==0.5.1
# pyasn1-modules==0.3.0
# pycparser==2.21
# pydantic==2.5.2
# pydantic_core==2.14.5
# Pygments==2.17.2
# PyJWT==2.8.0
# pymongo==4.6.1
# PyNaCl==1.5.0
# pyOpenSSL==23.2.0
# pyotp==2.9.0
# pyparsing==3.1.2
# pypng==0.20220715.0
# python-dateutil==2.9.0.post0
# python-engineio==4.9.0
# python-jose==3.3.0
# python-json-logger==2.0.7
python-socketio==5.11.1
# pytz==2023.4
# pywin32
# pywin32-ctypes==0.2.2
# pywinpty==2.0.12
# PyYAML==6.0.1
# pyzmq==25.1.0
# qrcode==7.4.2
# qtconsole==5.4.4
# QtPy==2.4.1
# referencing==0.30.2
# requests==2.31.0
# requests-oauthlib==1.4.0
# rfc3339-validator==0.1.4
# rfc3986-validator==0.1.1
# rich==13.7.1
# rpds-py==0.10.6
# rsa==4.9
# s3transfer==0.8.2
# SecretStorage==3.3.3
# Send2Trash==1.8.2
# service-identity==23.1.0
# shellingham==1.5.4
# simple-websocket==1.0.0
# six==1.16.0
# sniffio==1.3.0
# soupsieve==2.5
# speaklater3==1.4
# SQLAlchemy==2.0.28
# sqlparse==0.4.4
# sshtunnel==0.4.0
# stack-data==0.6.2
# storage==0.0.4.3
# terminado==0.17.1
# timeago==1.0.16
# tinycss2==1.2.1
# tornado==6.3.2
# tqdm==4.65.0
# traitlets==5.9.0
# Twisted==23.8.0
# twisted-iocpsupport==1.0.4
# txaio==23.1.1
# typer==0.9.0
# types-python-dateutil==2.8.19.14
# typing_extensions==4.10.0
# tzdata==2022.6
# ua-parser==0.18.0
# uri-template==1.3.0
# uritemplate==4.1.1
# urllib3==1.26.18
# user-agents==2.2.0
# wcwidth==0.2.6
# webcolors==1.13
# webencodings==0.5.1
websocket==0.2.1
websocket-client==1.6.4
Werkzeug==2.3.8
# widgetsnbextension==4.0.9
# wsproto==1.2.0
# WTForms==3.1.2
# yarl==1.9.2
# zipp==3.18.1
# zope.event==5.0
# zope.interface==6.0
